Conflicts:
	scripts/Makefile.am
	scripts/offload_slave_launch.sh.in
	src/arch/mpi/mpiremotenode.cpp
	src/arch/smp/smpplugin.cpp

+++ src/plugins/deps/cregions_deps.cpp
@@ -1,5 +1,5 @@
 /*************************************************************************************/
-/*      Copyright 2012 Barcelona Supercomputing Center                               */
+/*      Copyright 2015 Barcelona Supercomputing Center                               */
 /*                                                                                   */
 /*      This file is part of the NANOS++ library.                                    */
 /*                                                                                   */
@@ -300,15 +300,17 @@ namespace nanos {
                     TrackableObject &stat = *(*it);
                     TR1::unordered_map<TrackableObject*, bool>::iterator iterStat=statusMap.find (&stat);
                     if ( iterStat == statusMap.end() ){  
-                        if ( accessType.output ) {
-                            submitDependableObjectOutputNoWriteDataAccess( depObj, target, accessType, stat, callback );    
+                        if ( accessType.output && !accessType.concurrent && !accessType.commutative ) {
+                           submitDependableObjectOutputNoWriteDataAccess( depObj, target, accessType, stat, callback );    
                         } 
-                        submitDependableObjectInputNoReadDataAccess( depObj, target, accessType, stat, callback );  
+                        if ( accessType.input && !accessType.concurrent && !accessType.commutative ) {
+                           submitDependableObjectInputNoReadDataAccess( depObj, target, accessType, stat, callback );  
+                        }
                     } else {
                         bool isWriter=iterStat->second;
                         //This region was previously marked as "input" in this task, but our current writer 
                         //has to wait for it until all readers finish, reorder dependencies so we do not depend on ourselves
-                        if (!isWriter && accessType.output ) {          
+                        if (!isWriter && accessType.output && !accessType.concurrent && !accessType.commutative ) {
                             stat.getReadersLock().acquire();
                             stat.deleteReader(depObj);
                             stat.getReadersLock().release();
